Optimize `replace_params_with_values` (#13308)
* Add benchmark for `with_param_values`

Target of this benchmark: control that placeholders replacing does not get slower,
if the query does not contain placeholders at all.

* Do not save original expression name if it does not contain placeholders

Prior this patch, the original expression name was unconditionally preserved during the replacement of placeholders.
However, in some cases, this is unnecessary. If the expression does not contain any placeholders, its name remains unchanged, and we do not need to preserve it.

This patch adds the following optimization:

- If during first pass (placeholders type inference) was investigated that there are no placeholders, do not save the original name,
  do not apply second pass (replacing).
  Just leave the expression as it is and return `Transformed::no(...)`.