physical_plan: move metrics to task context
To share physical plans across executions we need to place
metrics in some other place. This patch moves them to the
task context. When plan is scanned metrics set is registered
in the context by the plan address.

To display a plan with metrics one should provide the task context,
where metrics associated with the plan are stored.

Also, fmt errors are fixed and applied several clippy suggestions.