dml: store destination as table source
This patch adds a field `dst` into struct `DmlStatement`. No reason to
do it in the other way than for a table scan, where we store the table
source. Whereas storing a dst adds the ability to analyze the table
for which DML operation is performed. It is useful for analyzer/optimizer
passes writing.